+++ src/brasil/gov/facetada/vocabulary.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+from Products.CMFCore.utils import getToolByName
 from eea.faceted.vocabularies.utils import IVocabularyFactory
 from zope.interface import implements
 from zope.schema.vocabulary import SimpleTerm
@@ -49,3 +50,25 @@ class TiposVocabulary(object):
         return SimpleVocabulary(items)
 
 TiposVocabularyFactory = TiposVocabulary()
+
+
+class ImageScaleVocabulary(object):
+    """ Create a vocabulary of image scales available in the site """
+    implements(IVocabularyFactory)
+
+    def __call__(self, context):
+        properties_tool = getToolByName(self, 'portal_properties')
+        imagescales_properties = getattr(properties_tool, 'imaging_properties', None)
+        raw_scales = getattr(imagescales_properties, 'allowed_sizes', None)
+
+        image_scales = {}
+        for line in raw_scales:
+            line = line.strip()
+            if line:
+                splits = line.split(' ')
+                if len(splits) == 2:
+                    name = line
+                    image_scales[name] = (splits[0], )
+        return SimpleVocabulary.fromValues(image_scales)
+
+ImageScaleVocabularyFactory = ImageScaleVocabulary()