+++ CHANGES.rst
@@ -4,5 +4,7 @@ Alterações
 1.0a1 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+  * Adiciona Widget de Banner.
+    [dbarbato]
   * Versão inicial do pacote.
     [dbarbato]